Pm stream

+++ PM.Core/PmInsufficientFileSizeException.cs
@@ -3,9 +3,9 @@
     public class PmInsufficientFileSizeException : ApplicationException
     {
         public PmInsufficientFileSizeException() : base() { }
-        public PmInsufficientFileSizeException(int originalSize) 
+        public PmInsufficientFileSizeException(long originalSize) 
             : base($"File size {originalSize} bytes is insufficient") { }
-        public PmInsufficientFileSizeException(int originalSize, int minimumSize) 
+        public PmInsufficientFileSizeException(long originalSize, long minimumSize) 
             : base($"File size {originalSize} bytes is insufficient, must be at least {minimumSize} bytes") { }
     }
 }
